refactor: extract leaderboard toggles
